% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractMeta.R
\name{extractMeta}
\alias{extractMeta}
\title{extractMeta}
\usage{
extractMeta(y)
}
\arguments{
\item{y}{A list of articles in a ProQuest text file already separated and pre-processed in prepare.R}
}
\value{
A data frame of document level meta-data variables and the full text in the variable text
}
\description{
extractMeta
}
