% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{prepare}
\alias{prepare}
\title{prepare}
\usage{
prepare(path = "path_name")
}
\arguments{
\item{path}{A character vector to the file location to be imported, processed with metadata extracted.}
}
\value{
A data.frame of document-level meta data and full text.
}
\description{
prepare
}
\examples{
prepare(path="inst/star.txt")
}
